Fix Slider arrows
