+++ src/pages/layout/Layout.js
@@ -38,14 +38,14 @@ const Layout = () => {
   // }
 
   return (
-    <div className="min-w-[380px] max-w-[1440px] mx-auto min-h-screen bg-white">
+    <div className="min-w-[380px] max-w-[1640px] mx-auto min-h-screen bg-white">
       <Sidebar
         sidebarMenuIsOpen={sidebarMenuIsOpen}
         toggleSidebarMenu={toggleSidebarMenu}
         breakpoint={MOBILE_BREAKPOINT}
       />
 
-      <div className={`duration-300 ${((windowWidth > MOBILE_BREAKPOINT) || sidebarMenuIsOpen) && 'ml-64'} `}>
+      <div className={`duration-300 ${((windowWidth > MOBILE_BREAKPOINT) || sidebarMenuIsOpen) && 'ml-[224px]'} `}>
         <Outlet />
       </div>
     </div>
